feat(schedule): improved function to get courses details
This commit adds the 'commentaire' field to the `final_dict` dictionary in the `build_final_dict` function. The value for this field is obtained from the `get_course_details` function. Additionally, a log message has been added to indicate when data is being written to a file in the `write_to_json` function.